refactor : LottoNumber 수정
* 근거 없는 long -> int 수정

+++ src/main/java/lotto/domain/LottoNumber.java
@@ -5,9 +5,9 @@ import java.util.Objects;
 public class LottoNumber implements Comparable<LottoNumber>{
     public static final int MINIMUM_LOTTO_NUMBER = 1;
     public static final int MAXIMUM_LOTTO_NUMBER = 45;
-    private long lottoNumber;
+    private int lottoNumber;
 
-    public LottoNumber(long lottoNumber) {
+    public LottoNumber(int lottoNumber) {
         if (lottoNumber < MINIMUM_LOTTO_NUMBER || lottoNumber > MAXIMUM_LOTTO_NUMBER) {
             throw new IllegalArgumentException("1 ~ 45 까지만 허용됩니다.");
         }
